// client/typescript/src/_generated/runtime/world/lineage/image.ts
export type ImageId = bigint;

export const ImageId = {
    /** Encode this value. */
    encode(writer: BinaryWriter, value: ImageId): void {
        encodeImageId(writer, value);
    },

    /** Decode one ImageId. */
    decode(reader: BinaryReader): ImageId {
        return decodeImageId(reader);
    },

    /** Return this value as JSON. */
    toJson(value: ImageId): Json {
        return toJsonImageId(value);
    },

    /** Return one ImageId from one JSON value. */
    fromJson(value: Json): ImageId {
        return fromJsonImageId(value);
    },
};